requirements.txt
numpy
pandas
matplotlib
seaborn
plotly
nbformat>=4.2.0
scikit-learn
category-encoders
missingno
ydata-profiling
setuptools
ipywidgets
xgboost
optuna
langchain==0.1.16
wikipedia
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
pydantic
anthropic
langchain-core
duckduckgo-search
fastapi
uvicorn
requests